+++ app/Http/Controllers/ProfileController.php
@@ -18,6 +18,8 @@ use Illuminate\Support\Facades\Auth;
 
 use Illuminate\Http\Request;
 
+use App\Policies\PostPolicy;
+
 class ProfileController extends Controller
 {
      public function show($user_id)
@@ -26,7 +28,7 @@ class ProfileController extends Controller
         $post = new Post;
         $user = User::find($user_id);
         $profile = Profile::where('user_id',$user_id)->first();
-        $post = Post::where('user_id',$user_id)->paginate(5);
+        $post = Post::where('user_id',$user_id)->orderBy('updated_at', 'DESC')->paginate(9);
         return view('profile.show')->with([
             'profile'=>$profile,
             'user'=>$user,
@@ -36,13 +38,16 @@ class ProfileController extends Controller
     
     public function myshow()
     {
-        $user_id=Auth::user()->id;
+        $like = new Like;
+        $user = Auth::user();
         $post = new Post;
-        $profile = Profile::where('user_id',$user_id)->first();
-        $post = Post::where('user_id',$user_id)->paginate(5);
+        $profile = Profile::where('user_id',$user->id)->first();
+        $post = Post::where('user_id',$user->id)->orderBy('updated_at', 'DESC')->paginate(9);
         return view('profile.mypage')->with([
             'profile'=>$profile,
-            'posts'=>$post]);
+            'posts'=>$post,
+            'user'=>$user,
+            'like_model'=>$like]);
         
     }
      public function create()
@@ -64,20 +69,29 @@ class ProfileController extends Controller
             $profile->save();
             }
             
-        return redirect('/profiles/'.$profile->user_id);
+        return redirect('/profiles/mypage');
         
     }
     
     public function edit(Profile $profile)
     {
+        $user=Auth::user();
+        $this->authorize('view', $profile);
+        
         return view('profile.edit')->with(['profile'=>$profile]);
     }
     
      public function update(ProfileRequest $request, Profile $profile)
     {
-
-        $input=$request['profile.body'];
-        $profile->body=$input;
+        $user=Auth::user();
+        $this->authorize('update', $profile);
+        
+        $input_name=$request['user.name'];
+        $user = User::find($profile->user_id);
+        $user->name=$input_name;
+        $user->save();
+        $input_body=$request['profile.body'];
+        $profile->body=$input_body;
         if($request['profile.image_path']==null){
              $profile->save();
         }else{
@@ -86,16 +100,20 @@ class ProfileController extends Controller
             $profile->save();
             }
         
-        return redirect('/profiles/'.$profile->user_id);
+        return redirect('/profiles/mypage');
     
     }
     
     public function destroy(Profile $profile,Request $request)
     {
+        $user=Auth::user();
+        $this->authorize('delete', $profile);
+        
         Storage::disk('s3')->delete(parse_url($profile->image_path,PHP_URL_PATH));
         $profile->image_path=null;
         $profile->save();
-        return redirect('/profiles/'.$profile->user_id);
+        
+        return redirect('/profiles/'.$profile->id.'/edit');
     }
     
 }
